// ui/src/features/dashboards/api/dashboard-detail.generated.ts
type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
// ...
export type DashboardDetailQueryVariables = Exact<{
  id: string
}>

export type DashboardDetailQuery = {
  readonly metricNames: ReadonlyArray<string>
  readonly dashboard: {
    readonly id: string
    readonly name: string
    readonly layout: string
  } | null
}

export const DashboardDetailDocument = {
  kind: "Document",
  definitions: [
    {
      kind: "OperationDefinition",
      operation: "query",
      name: { kind: "Name", value: "DashboardDetail" },
      variableDefinitions: [
        {
          kind: "VariableDefinition",
          variable: { kind: "Variable", name: { kind: "Name", value: "id" } },
          type: {
            kind: "NonNullType",
            type: {
              kind: "NamedType",
              name: { kind: "Name", value: "String" },
            },
          },
        },
      ],
      selectionSet: {
        kind: "SelectionSet",
        selections: [
          {
            kind: "Field",
            name: { kind: "Name", value: "dashboard" },
            arguments: [
              {
                kind: "Argument",
                name: { kind: "Name", value: "id" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "id" },
                },
              },
            ],
            selectionSet: {
              kind: "SelectionSet",
              selections: [
                { kind: "Field", name: { kind: "Name", value: "id" } },
                { kind: "Field", name: { kind: "Name", value: "name" } },
                { kind: "Field", name: { kind: "Name", value: "layout" } },
              ],
            },
          },
          { kind: "Field", name: { kind: "Name", value: "metricNames" } },
        ],
      },
    },
  ],
} as unknown as DocumentNode<DashboardDetailQuery, DashboardDetailQueryVariables>
// ...
export const DashboardDetailQuerySchema: z.ZodType<DashboardDetailQuery> = z.object({
  dashboard: z
    .object({
      id: z.string(),
      name: z.string(),
      layout: z.string(),
    })
    .nullable(),
  metricNames: z.array(z.string()),
})