// ui/src/features/dashboards/api/dashboard-save.generated.ts
type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
// ...
export type DashboardSaveMutationVariables = Exact<{
  name: string
  layout: string
  id?: string | null | undefined
}>

export type DashboardSaveMutation = {
  readonly dashboardSave: {
    readonly id: string
    readonly name: string
    readonly layout: string
    readonly updatedAtNanos: string
  }
}

export const DashboardSaveDocument = {
  kind: "Document",
  definitions: [
    {
      kind: "OperationDefinition",
      operation: "mutation",
      name: { kind: "Name", value: "DashboardSave" },
      variableDefinitions: [
        {
          kind: "VariableDefinition",
          variable: { kind: "Variable", name: { kind: "Name", value: "name" } },
          type: {
            kind: "NonNullType",
            type: {
              kind: "NamedType",
              name: { kind: "Name", value: "String" },
            },
          },
        },
        {
          kind: "VariableDefinition",
          variable: {
            kind: "Variable",
            name: { kind: "Name", value: "layout" },
          },
          type: {
            kind: "NonNullType",
            type: {
              kind: "NamedType",
              name: { kind: "Name", value: "String" },
            },
          },
        },
        {
          kind: "VariableDefinition",
          variable: { kind: "Variable", name: { kind: "Name", value: "id" } },
          type: { kind: "NamedType", name: { kind: "Name", value: "String" } },
        },
      ],
      selectionSet: {
        kind: "SelectionSet",
        selections: [
          {
            kind: "Field",
            name: { kind: "Name", value: "dashboardSave" },
            arguments: [
              {
                kind: "Argument",
                name: { kind: "Name", value: "name" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "name" },
                },
              },
              {
                kind: "Argument",
                name: { kind: "Name", value: "layout" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "layout" },
                },
              },
              {
                kind: "Argument",
                name: { kind: "Name", value: "id" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "id" },
                },
              },
            ],
            selectionSet: {
              kind: "SelectionSet",
              selections: [
                { kind: "Field", name: { kind: "Name", value: "id" } },
                { kind: "Field", name: { kind: "Name", value: "name" } },
                { kind: "Field", name: { kind: "Name", value: "layout" } },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "updatedAtNanos" },
                },
              ],
            },
          },
        ],
      },
    },
  ],
} as unknown as DocumentNode<DashboardSaveMutation, DashboardSaveMutationVariables>
// ...
export const DashboardSaveMutationSchema: z.ZodType<DashboardSaveMutation> = z.object({
  dashboardSave: z.object({
    id: z.string(),
    name: z.string(),
    layout: z.string(),
    updatedAtNanos: z.string(),
  }),
})