// ui/src/features/dashboards/api/dashboards-list.generated.ts
type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
// ...
export type DashboardsListQueryVariables = Exact<{ [key: string]: never }>

export type DashboardsListQuery = {
  readonly metricNames: ReadonlyArray<string>
  readonly dashboards: ReadonlyArray<{
    readonly id: string
    readonly name: string
    readonly layout: string
    readonly updatedAtNanos: string
  }>
}

export const DashboardsListDocument = {
  kind: "Document",
  definitions: [
    {
      kind: "OperationDefinition",
      operation: "query",
      name: { kind: "Name", value: "DashboardsList" },
      selectionSet: {
        kind: "SelectionSet",
        selections: [
          {
            kind: "Field",
            name: { kind: "Name", value: "dashboards" },
            selectionSet: {
              kind: "SelectionSet",
              selections: [
                { kind: "Field", name: { kind: "Name", value: "id" } },
                { kind: "Field", name: { kind: "Name", value: "name" } },
                { kind: "Field", name: { kind: "Name", value: "layout" } },
                { kind: "Field", name: { kind: "Name", value: "updatedAtNanos" } },
              ],
            },
          },
          { kind: "Field", name: { kind: "Name", value: "metricNames" } },
        ],
      },
    },
  ],
} as unknown as DocumentNode<DashboardsListQuery, DashboardsListQueryVariables>
// ...
export const DashboardsListQuerySchema: z.ZodType<DashboardsListQuery> = z.object({
  dashboards: z.array(
    z.object({
      id: z.string(),
      name: z.string(),
      layout: z.string(),
      updatedAtNanos: z.string(),
    })
  ),
  metricNames: z.array(z.string()),
})